+++ verik-plugin/src/main/kotlin/io/verik/plugin/config/AuroraTargetConfig.kt
@@ -1,0 +1,19 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.verik.plugin.config
+
+import java.nio.file.Path
+
+/**
+ * Configuration for an aurora target.
+ */
+class AuroraTargetConfig(
+    override val toolchain: String,
+    override val timestamp: String,
+    override val projectName: String,
+    override val targetName: String,
+    val compileTops: List<String>,
+    val dpiLibs: List<Path>
+) : TargetConfig()